cmd/internal/gc: inline non-escaping convT2*
This is a follow-up to CL 8201.

Non-escaping convT2* can now be entirely inlined.
Such calls are currently rare (only three in the stdlib),
but as escape analysis improves there will be more of them.

As an immediate benefit, the number of arguments
to convT2* goes back down by one. This speeds up
and shrinks all convT2* call sites.
The godoc binary shrinks by ~25k with this CL.

While we're here, remove handling of direct ifaces from convT2*;
it is unreachable.

Change-Id: I3f49482dc8db71c3a5657d1d1edf3d77f0487fc1