cmd/internal/gc: diagnose missed registerization opportunities
CL 91850043 significantly improved
registerization in 6g and 8g.
Early versions of the CL included diagnostic code,
but it was removed for submission.

Add a new debug flag and revive the diagnostics,
so that they can be run easily.
This will also allow bug fixes in this arena
to be locked in with tests.

Change-Id: I19c9d6ddd40c72b9b22c01bfdeabd0c365d2550c