cmd/compile: make algs content-addressable
Name alg functions after the type structure
rather than the type name.
This allows structurally equivalent types
to share algs.

Shrinks cmd/go by 105k (-1.2%).

DO NOT SUBMIT

This is marked as DO NOT SUBMIT because
I am not 100% sure that this is safe,
and I'm not sure what steps need to be taken
to make it so. Input requested.

The concern is the stack map for the generated functions.
For example, both int64 and *byte are 64-bit memory comparable.
However, the are not the same from a stackmap perspective.
So the generated code for

type S struct {
	a       int64
	b       byte
	c, d, e int64
}

should perhaps be different than for

type P struct {
	a       int64
	b       byte
	c, d, e *int64
}

but as it stands, they will share implementations ({0m9,16m24}).

Change-Id: I77f42a669267c3bb52be0e937d9ae6b163297253