[dev.ssa] cmd/compile: use lookup table for ssa rewrites
DO NOT REVIEW

This is an experiment, less interesting in itself
than for the general question of compiling such 
switch statements. The compiler could
easily recognize this form and make the transformation.

However, this increases binary size non-trivially.
The static lookup table is smaller than the
corresponding switch function, and (in theory)
nothing else has changed.
So why does the cmd/compile grow?

Change-Id: I1f94b2044af7b6d4c098c1ae11f006a3eb41c990