cmd/compile: add some ssa rule debugging support
DO NOT SUBMIT

This is a tool I've been using to debug why an SSA rule isn't firing.
To use, add "DEBUG " before an SSA rule and regenerate, e.g.

DEBUG (Mul8  (Const8  [-1]) x) -> (Neg8  x)

The compiler will print each attempt to match the rule and why it failed.
Not all failure paths are supported yet, nor are rules for blocks.

Marked as DNS because the implementation is incomplete
and the usage is clunky:

* it can be way too noisy, with no good way to filter (perhaps needs to print line numbers? accept a condition to check?)
* needs to be easy to turn on/off (envvar?) for any given compile, defaulting to off
* the DEBUG prefix feels like an ugly hack, maybe something like
an OpDebug would be better and somehow useful elsewhere (?)

I'd be delighted if anyone wanted to fix this up
and make it permanent.

Change-Id: Iceb1cdd5580368cbfdf62252449beddbc6c64d04