strconv: add and use 32 bit formatBits routine
DO NOT SUBMIT

[This is a preliminary CL for discussion.]

Most numbers are small.
This CL adds a 32 bit formatBits routine
and switches to it for numbers that fit in 32 bits.

I was surprised by the impact.
(The FormatInt number is noise. It is included
to show that the extra branch is not too expensive.)

name              old time/op  new time/op  delta
FormatInt-8       1.97µs ± 1%  1.95µs ± 2%   -0.68%  (p=0.026 n=20+20)
FormatIntSmall-8   511ns ± 2%   457ns ± 2%  -10.68%  (p=0.000 n=20+20)

Questions and TODOs:

* Is the code weight worth 10% performance?

* Can formatBits32 be optimized more?
  All I did was copy/paste formatBits and change obvious things.

* Would formatBits16 or formatBits8 be worth it?

* Use formatBits32 in other exported functions,
  or have formatBits do the 32 bit test and divert.

* Does it move other, higher level benchmarks?

* Is a similar trick worth doing in package fmt,
  which IIRC has its own number formatting routines?

* Does anyone else want to see this through? :)
  I'd like to stay focused on the compiler at the moment.

Change-Id: I244ecb7ef22dcef673594a288ef1e027f5fb61b5