cmd/internal/obj: TLSBSS is only for globals
Refuse to mark functions as TLSBSS,
and only check globals for the TLSBSS
flag in the arm64 compiler.
This avoids a data race on a.Sym.Type,
in which a function is referred to by its Sym,
thus requiring a read of its Sym.Type,
which can happen concurrently with the Sym.Type
being written to in Flushplist as part of compiling
the called function.

While we're here, eliminate package log in favor of ctxt.Diag.

Updates #15756

Change-Id: I1768fd82811b4ece09c492a419288d898730325f