cmd/go, cmd/compile: don't emit compiler info for executable package
Shaves a percent off the time to do an incremental
compile+link of cmd/go:

RebuildCmdGo   629ms ± 4%   621ms ± 6%  -1.23%  (p=0.000 n=82+100)

Maybe not worth it, since it's just a percent,
but it is a common case.

Special care has to be taken to ensure that
importing package main (e.g. in tests)
continues to work. See #4210.

Fortunately, we do that a lot in the std tests,
so this has plenty of test coverage.

Change-Id: I449648fb0758f338585544c7b45bf181e69a6560