cmd/compile: indicate type-by-type in binary export whether it was recorded
DO NOT SUBMIT

[for discussion]

Currently there are two modes for recording types:
named types or all types.
The latter is disabled due to bugs.

This makes it hard to incrementally fix these bugs,
to incrementally record safe types, and to experiment.

Instead, indicate with every type tag whether it was
recorded or not. This has no binary size cost.

Change-Id: I3af19e5b88b848b00815ea9d55648c31431c88f0