cmd/compile: pass an hmap to makemap when initializing static maps
DO NOT SUBMIT

[I am not sure the best way to identify whether
a particular OMAPLIT is safe for a static hmap;
see the TODO in the CL]

Prior to this CL, we did code generation for a map literal like this:

m := make(maptype, number of elements)
m[k] = v // for all elements, sometimes using a loop

The assignment to m generates a write barrier.
Each write barrier generates a new SSA block.
In giant composite literals, this can lead to lots of blocks.
This led to a panic when compiling the code in #19751.

This CL changes code generation when compiling a static
composite literal to do:

var stat hmap // static symbol
hp := &stat
m := (map)(hp)
initmap(hp, number of elements)
m[k] = v

This involves no assignments, and thus no write barriers.
As a result, the code in #19751 compiles.
Slowly still, but the compilation completes.

Change-Id: Ic0c05f5892e50fadc5e6d6adb5a3f5275ec38224