runtime: improve BenchmarkMapPopulate
BenchmarkMapPopulate measures the cost of populating a map.
However, there are two interesting cases: Without hint, so the map has to grow,
and with hint, so there's no growth.

Support both.

At the same time, reduce the range of the map size.
Size 1 really just measures map creation, for which we have other benchmarks,
and the largest size doesn't add much value in practice.

As a result, the number of benchmark runs only grows from 4 to 6.

Change-Id: I12deb26cfc7212822365efd98f97bec496fa10fa