cmd/compile: disable magic division during first opt pass
Magic division disrupts our ability to detect
obvious boundededness during the opt pass.
Usually we avoid order dependence by writing
rewrite rules to incorporate the work done by other rules.
In the case of magic division, that is infeasible;
the optimized form is too hard to detect and invert.
Instead, defer magic division to the second opt pass.
This is ugly, but it appears to work well in practice.

Change-Id: I0634477ebe20b9afb9a0e7a3b65d6786d97f5268