cmd/compile: ensure alg functions are generated even without DWARF
We were relying on calls to ngotype from within DWARF generation
to add types to the signats list, from which we generate algs.
When DWARF was disabled, some alg functions were not generated.

Any time we know we might generate calls to eq or hash,
ensure that the types are on the signats list.

This is based on CL 122403 by Cherry Zhang,
but with the intent of being a more narrowly targeted fix.

Fixes #23546.

Change-Id: I8a497a07ad7a4a072cb03700f83b90a9accf2fd0