cmd/compile: combine if statements in generated rules
DO NOT SUBMIT

I do not like the approach taken in this CL;
it it too hard to read and reason about the rulegen code.

A better approach would be to build a data structure,
optimize it a bit, and then print it,
rather than working off a textual representation.

I'm mailing it as a demonstration of what the benefits of
such an approach might be. It makes the generated code
a lot shorter and more readable, IMHO.

Change-Id: I3303361f1bfe0345967a499b4c2f3a771356efea