cmd/compile: include deadcode and cse in rewrite-until-stable loop
DO NOT SUBMIT

This change doesn't pull its weight.

----

Some of the rewrite rules are context dependent,
in a way that deadcode and CSE affects,
particularly the merged load/store operations on amd64.
For example, if a rule depends on v.Uses == 1,
and all but one uses are dead but still present,
running deadcode will enable that rule to fire.

This change increases the effectiveness of the other
changes in the CL series.
In fact, without it, the ADCQload* and SBBQload*
rules do not fire at all.

Another option would be to add another rules pass.
We would do lower, then lowered CSE, then deadcode,
then the second rules pass.
However, some rules would have to be duplicated
in both passes, particularly the rules to absorb
InvertFlags into blocks and other values.
This change follows the spirit of the rewrite
rules better, and avoids duplication.

Change-Id: I45e57a9e68c6e9452e3993c1250c1fe4a7d2527f