cmd/compile: include deadcode in rewrite-until-stable loop
Some of the rewrite rules are context dependent,
in a way that deadcode affects.
For example, if a rule depends on v.Uses == 1,
and all but one uses are dead but still present,
running deadcode will enable that rule to fire.

This change adds a call to deadcode inside optimization rewrite loops.

We must take care not to call deadcode during decomposition rewrites.
This is because they rely on the existence of dead values
to correctly identify named values.