cmd/compile: teach flagalloc to handle flag-memory-flag Values
DO NOT SUBMIT

This is not needed unless we add ADCQload*
and/or SBBQload*, and I don't believe we will add them.

---

Flagalloc knew how to handle flag-generating values
that had a memory argument.
It did not correctly handle flag-generating values
that had a memory argument *and* a flag argument.
Spilling such a value requires recursively spilling
everything else that it depends on as well.

It also requires looking through OpSelectN values to the
underlying value when planning what will be spilled.
Since values have already been scheduled, flagalloc
must be careful to generate all args a value needs
before the value itself.
If you don't look through OpSelectN values,
then the underlying value has been issued
before you have a chance to regenerate the
arguments that it depends on.

None of this mattered in the past, because there were no
flag-generating, memory-dependent, flag-dependent values.
A subsequent CL will introduce some.

Change-Id: Ib86909f291f785957ac30c5006eb959e494c5a32