unicode/utf8: optimize checking against an acceptRange
An acceptRange is used to check whether a byte falls in a range.
The existing code used the naive way: lo <= x <= hi.
Because bytes are unsigned, this can be rewritten as x-lo <= hi-lo.

The compiler does this optimization (modulo #30645),
but only when lo and hi are constants.
In this case, lo and hi are effectively constants,
but the compiler can't see that.

Change-Id: I9a4a87393915168bd59e28c8048e85e4c688aca5